requirements.txt
bs4==4.12.2
docx2txt==0.8
lxml==4.9.3
python-docx==1.0.1
requests==2.31.0
telebot==4.14.0
